refactor: remove insta style post from community section and add it as custom component for future
